+++ cvxpy/atoms/affine/sum_entries.py
@@ -18,11 +18,12 @@ along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 """
 
 from cvxpy.atoms.affine.affine_atom import AffAtom
+from cvxpy.atoms.axis_atom import AxisAtom
 import cvxpy.utilities as u
 import cvxpy.lin_ops.lin_utils as lu
 import numpy as np
 
-class sum_entries(AffAtom):
+class sum_entries(AxisAtom, AffAtom):
     """ Summing the entries of an expression.
 
     Attributes
@@ -31,19 +32,14 @@ class sum_entries(AffAtom):
         The expression to sum the entries of.
     """
 
-    def __init__(self, expr):
-        super(sum_entries, self).__init__(expr)
+    def __init__(self, expr, axis=None):
+        super(sum_entries, self).__init__(expr, axis=axis)
 
     @AffAtom.numpy_numeric
     def numeric(self, values):
         """Sums the entries of value.
         """
-        return np.sum(values[0])
-
-    def shape_from_args(self):
-        """Always scalar.
-        """
-        return u.Shape(1, 1)
+        return np.sum(values[0], axis=self.axis)
 
     @staticmethod
     def graph_implementation(arg_objs, size, data=None):
@@ -63,4 +59,16 @@ class sum_entries(AffAtom):
         tuple
             (LinOp for objective, list of constraints)
         """
-        return (lu.sum_entries(arg_objs[0]), [])
+        axis = data[0]
+        if axis is None:
+            obj = lu.sum_entries(arg_objs[0])
+        elif axis == 1:
+            const_size = (arg_objs[0].size[1], 1)
+            ones = lu.create_const(np.ones(const_size), const_size)
+            obj = lu.rmul_expr(arg_objs[0], ones, size)
+        else: # axis == 0
+            const_size = (1, arg_objs[0].size[0])
+            ones = lu.create_const(np.ones(const_size), const_size)
+            obj = lu.mul_expr(ones, arg_objs[0], size)
+
+        return (obj, [])